+++ plone/app/content/interfaces.py
@@ -35,22 +35,3 @@ class IIndexableObjectWrapper(Interface):
         """Update the wrapper with variables from e.g. the workflow
         tool.
         """
-
-class ISiteContentSettings(Interface):
-    """ Settings in the registry which determine various admin configurable global behaviour of content items
-    """
-    file_mimetype_behaviour = schema.Dict(
-        title=u"File item behaviour per content type (can use '*')",
-        key_type=schema.ASCIILine(title=u"Mimetype"),
-        value_type=schema.Choice(
-            values=("inline","attachment","view"),
-            title=u"Behaviour",
-            ),
-        default={'application/msword':'inline',
-                 'application/x-msexcel':'inline',
-                 'application/vnd.ms-excel':'inline',
-                 'application/vnd.ms-powerpoint':'inline',
-                 'application/pdf':'inline',
-                 'application/x-shockwave-flash':'inline',
-                 '*':'attachment'}
-    )